updates by Skala
merged:

  gentoo: f9b418beb73f9201ad906961898b0217ccef904f
  party-overlay: 55973e13049c60e011ecbc7a20edda5be5922d2a
  ubuntu-overlay: dd2f24623aac9a50bba380a5c8ddbb95b5ffab04
  stable-overlay: 0819ea53d72d7451432a285001942ee720371d71
  elementary: 8b20be0bad920b04901bdd49a2bd7ef5becd3566
  funtoo-overlay: dd76fcaaebc70b5de863f6395239a53e9e9d8dae
  sabayon-for-gentoo: d009cc1016c433f66302c7b973fc894e7c332f12
  sabayon-tools: 471ff9eb037189aca0a97d664d96872d9a38dab9
  gamerlay: 4bbdd8768371bef56d8c9b0b738493673b535d5d
  unity-gentoo: e55a5050ec680d489bec79403c4317361c2bbac1
  gnome-next: 604a1517315e30bf6c5b8505e95542175926cb67
  flatpak-overlay: dba2292fa92fbe7f51a437ff89d27cbab6105b67

+++ app-admin/sysklogd/sysklogd-2.0.ebuild
@@ -12,11 +12,11 @@ SRC_URI="https://github.com/troglobit/sysklogd/releases/download/v$(ver_cut 1-2)
 LICENSE="BSD"
 SLOT="0"
 KEYWORDS="~alpha ~amd64 ~arm ~arm64 ~hppa ~ia64 ~mips ~ppc ~ppc64 ~s390 ~sh ~sparc ~x86"
-IUSE="klogd logrotate systemd"
+IUSE="klogd logger logrotate systemd"
 RESTRICT="test"
 
-DEPEND=""
-RDEPEND=""
+DEPEND="!sys-apps/util-linux[logger]"
+RDEPEND="${DEPEND}"
 
 DOCS=( ChangeLog.md README.md )
 
@@ -36,9 +36,8 @@ src_prepare() {
 
 src_configure() {
 	local myeconfargs=(
-		# we have logger from sys-apps/util-linux
-		--without-logger
 		$(use_with klogd)
+		$(use_with logger)
 		$(use_with systemd systemd $(systemd_get_systemunitdir))
 	)
 	econf "${myeconfargs[@]}"
@@ -51,8 +50,22 @@ src_install() {
 	doins syslog.conf
 	keepdir /etc/syslog.d
 
-	find "${ED}" -type f \( -name "*.a" -o -name "*.la" \) -delete || die
-
 	newinitd "${FILESDIR}"/sysklogd.rc8 sysklogd
-	newconfd "${FILESDIR}"/sysklogd.confd sysklogd
+	newconfd "${FILESDIR}"/sysklogd.confd2 sysklogd
+
+	if use logrotate ; then
+		insinto /etc/logrotate.d
+		newins "${FILESDIR}"/sysklogd.logrotate sysklogd
+		sed 's@ -r 10M:10@@' -i "${ED}"/etc/conf.d/sysklogd || die
+	fi
+
+	find "${ED}" -type f \( -name "*.a" -o -name "*.la" \) -delete || die
+}
+
+pkg_postinst() {
+	if ! use logrotate && [[ -n ${REPLACING_VERSIONS} ]] && ver_test ${REPLACING_VERSIONS} -lt 2.0 ; then
+		elog "Starting with version 2.0 syslogd has built in log rotation"
+		elog "functionality that does no longer require a running cron daemon."
+		elog "So we no longer install any log rotation cron files for sysklogd."
+	fi
 }
